requirements.txt
# Core Data
polars>=1.0.0
pyarrow>=15.0.0

# Scraping
requests>=2.31.0
beautifulsoup4>=4.12.0
playwright>=1.40.0
tenacity>=8.2.0
lxml>=5.0.0

# Machine Learning
xgboost>=2.0.0
scikit-learn>=1.4.0
optuna>=3.5.0
shap>=0.45.0

# Experiment Tracking
mlflow>=2.10.0

# Data Validation
pandera>=0.18.0

# API
fastapi>=0.110.0
uvicorn>=0.27.0
pydantic>=2.5.0

# Database
psycopg2-binary>=2.9.0

# Monitoring
streamlit>=1.30.0
plotly>=5.18.0

# Scheduling
schedule>=1.2.0

# Configuration
pyyaml>=6.0.0

# Testing
pytest>=8.0.0
pytest-cov>=4.1.0
pytest-asyncio>=0.23.0
